requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
werkzeug==2.2.3
pymysql==1.0.3